# lib/api-zod/src/generated/types/recurringMistake.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Frenchami learning API
 * OpenAPI spec version: 0.1.0
 */

export interface RecurringMistake {
  pattern: string;
  explanation: string;
  count: number;
  lastSeenAt: Date;
}