# lib/api-zod/src/generated/types/tutorHistoryMessage.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Frenchami learning API
 * OpenAPI spec version: 0.1.0
 */
import type { TutorHistoryMessageRole } from './tutorHistoryMessageRole';

export interface TutorHistoryMessage {
  role: TutorHistoryMessageRole;
  /**
     * @minLength 1
     * @maxLength 2000
     */
  content: string;
}