# lib/api-zod/src/generated/types/tutorMessageInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Frenchami learning API
 * OpenAPI spec version: 0.1.0
 */
import type { TutorHistoryMessage } from './tutorHistoryMessage';
import type { TutorMessageInputLevel } from './tutorMessageInputLevel';

export interface TutorMessageInput {
  level: TutorMessageInputLevel;
  /**
     * @minLength 1
     * @maxLength 2000
     */
  message: string;
  /** @maxItems 20 */
  history: TutorHistoryMessage[];
}